// frontend/sdk/models/GauntletHistoryEntryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Unpaused Spotify API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TrackEntity } from './TrackEntity';
import {
    TrackEntityFromJSON,
    TrackEntityFromJSONTyped,
    TrackEntityToJSON,
    TrackEntityToJSONTyped,
} from './TrackEntity';

/**
 * 
 * @export
 * @interface GauntletHistoryEntryDto
 */
export interface GauntletHistoryEntryDto {
    /**
     * The gauntlet run ID
     * @type {string}
     * @memberof GauntletHistoryEntryDto
     */
    id: string;
    /**
     * Date of the run (yyyy-MM-dd)
     * @type {string}
     * @memberof GauntletHistoryEntryDto
     */
    date: string;
    /**
     * Final score (consecutive correct guesses)
     * @type {number}
     * @memberof GauntletHistoryEntryDto
     */
    score: number;
    /**
     * Difficulty of the run
     * @type {string}
     * @memberof GauntletHistoryEntryDto
     */
    difficulty: GauntletHistoryEntryDtoDifficultyEnum;
    /**
     * The tracks that were played in the run
     * @type {Array<TrackEntity>}
     * @memberof GauntletHistoryEntryDto
     */
    tracks: Array<TrackEntity>;
    /**
     * How many seconds the run lasted
     * @type {number}
     * @memberof GauntletHistoryEntryDto
     */
    durationSeconds: number;
}


/**
 * @export
 */
export const GauntletHistoryEntryDtoDifficultyEnum = {
    Easy: 'EASY',
    Medium: 'MEDIUM',
    Hard: 'HARD',
    Expert: 'EXPERT'
} as const;
export type GauntletHistoryEntryDtoDifficultyEnum = typeof GauntletHistoryEntryDtoDifficultyEnum[keyof typeof GauntletHistoryEntryDtoDifficultyEnum];


/**
 * Check if a given object implements the GauntletHistoryEntryDto interface.
 */
export function instanceOfGauntletHistoryEntryDto(value: object): value is GauntletHistoryEntryDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('date' in value) || value['date'] === undefined) return false;
    if (!('score' in value) || value['score'] === undefined) return false;
    if (!('difficulty' in value) || value['difficulty'] === undefined) return false;
    if (!('tracks' in value) || value['tracks'] === undefined) return false;
    if (!('durationSeconds' in value) || value['durationSeconds'] === undefined) return false;
    return true;
}

export function GauntletHistoryEntryDtoFromJSON(json: any): GauntletHistoryEntryDto {
    return GauntletHistoryEntryDtoFromJSONTyped(json, false);
}

export function GauntletHistoryEntryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): GauntletHistoryEntryDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'date': json['date'],
        'score': json['score'],
        'difficulty': json['difficulty'],
        'tracks': ((json['tracks'] as Array<any>).map(TrackEntityFromJSON)),
        'durationSeconds': json['durationSeconds'],
    };
}

export function GauntletHistoryEntryDtoToJSON(json: any): GauntletHistoryEntryDto {
    return GauntletHistoryEntryDtoToJSONTyped(json, false);
}

export function GauntletHistoryEntryDtoToJSONTyped(value?: GauntletHistoryEntryDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'date': value['date'],
        'score': value['score'],
        'difficulty': value['difficulty'],
        'tracks': ((value['tracks'] as Array<any>).map(TrackEntityToJSON)),
        'durationSeconds': value['durationSeconds'],
    };
}